Extract snackbar command to the main controller
