+++ webapp/lib/controller.dart
@@ -7,6 +7,7 @@ import 'package:katikati_ui_lib/components/auth/auth.dart';
 import 'package:katikati_ui_lib/components/logger.dart';
 
 import 'package:katikati_ui_lib/components/model/model.dart' as model;
+import 'package:katikati_ui_lib/components/snackbar/snackbar.dart';
 import 'package:nook/platform/platform.dart';
 import 'view.dart';
 
@@ -45,6 +46,9 @@ enum BaseAction {
 
   signInButtonClicked,
   signOutButtonClicked,
+
+  showSnackbar,
+  showBanner,
 }
 
 class Data {}
@@ -74,6 +78,24 @@ class SystemMessagesData extends Data {
   @override
   String toString() => 'SystemMessagesData: {messages: ${messages.map((m) => m.toData().toString())}}';
 }
+
+class SnackbarData extends Data {
+  String text;
+  SnackbarNotificationType type;
+  SnackbarData(this.text, this.type);
+
+  @override
+  String toString() => 'SnackbarData: {text: $text, type: $type}';
+}
+
+class BannerData extends Data {
+  String text;
+  BannerData(this.text);
+
+  @override
+  String toString() => 'BannerData: {text: $text}';
+}
+
 
 class Controller {
   model.User signedInUser;
@@ -135,6 +157,16 @@ class Controller {
         } else {
           view.bannerView.hideBanner();
         }
+        break;
+
+      case BaseAction.showSnackbar:
+        SnackbarData snackbarData = data;
+        view.snackbarView.showSnackbar(snackbarData.text, snackbarData.type);
+        break;
+
+      case BaseAction.showBanner:
+        BannerData bannerData = data;
+        view.bannerView.showBanner(bannerData.text);
         break;
       default:
     }
